refactor(filid): split findSubtreeUsages below the CC threshold
The organ-BFS fix pushed findSubtreeUsages to CC 16 (> 15) — caught by
this branch's own re-review evidence pass. Extract the subtree walk to
collectSubtreeNodes.ts and the per-file import check to a private helper;
behaviour unchanged, organ-traversal regression test still green.
